refactor: extract AppUtilityActionResultPresenter.swift from AppUtilityActionController.swift
Splits the `AppUtilityActionResultPresenter` @MainActor class (~48
lines) out of AppUtilityActionController.swift into its own file so
that file holds only the result enum + the controller.

Byte-identical move of lines 9-55. Both files import AppKit + Foundation.

AppUtilityActionController.swift: 161 → 112 lines. Tests: 667.
